mysite_latest/requirements.txt
Django==2.1.5
django-allauth==0.39.1
django-anymail==6.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
gunicorn==19.9.0
psycopg2==2.7.7
PyJWT==1.7.1
requests==2.21.0
requests-oauthlib==1.2.0
selenium==2.53.6
whitenoise==4.1.2